Fix ModelServerFactory empty register and rename cv/lm model folders
