+++ uniflow/flow/extract/extract_pdf_flow.py
@@ -8,7 +8,7 @@ from uniflow.node import Node
 from uniflow.op.extract.load.pdf_op import ExtractPDFOp, ProcessPDFOp
 from uniflow.op.extract.split.constants import PARAGRAPH_SPLITTER
 from uniflow.op.extract.split.splitter_factory import SplitterOpsFactory
-from uniflow.op.model.CV.cv_model import CvModel
+from uniflow.op.model.cv.model import CvModel
 
 
 class ExtractPDFFlow(Flow):
