+++ uniflow/op/model/cv/__init__.py
@@ -1,0 +1,1 @@
+from uniflow.op.model.cv.model_server import *  # noqa: F401, F403
