+++ uniflow/op/model/cv/model_server.py
@@ -181,7 +181,7 @@ class LayoutModelServer(AbsModelServer):
         import cv2  # pylint: disable=import-outside-toplevel
         import numpy as np  # pylint: disable=import-outside-toplevel
 
-        from uniflow.op.model.CV.layout_utils import (  # pylint: disable=import-outside-toplevel
+        from uniflow.op.model.cv.layout_utils import (  # pylint: disable=import-outside-toplevel
             XYCut,
         )
